server/requirements.txt
Django==1.10.5
django-cors-middleware==1.3.1
django-extensions==1.7.1
djangorestframework==3.4.4
PyJWT==1.4.2
six==1.10.0
python-dotenv==0.7.1
sendgrid-django==4.2.0
stripe==1.77.2
